# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Radware Vulnerable E-commerce API",
    "version": "1.2.0",
    "description": "An intentionally vulnerable e-commerce API designed to demonstrate business logic attacks, focusing on path and query parameters. This version includes Protected Entity Mode which blocks destructive actions on certain demo entities.\n"
  },
...
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List all products",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "description": "Filter products based on user role (e.g., 'user'). This parameter seems to be sent by the UI.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "admin",
                "guest"
              ],
              "default": "user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a new product (BFLA Target)",
        "description": "Creates a new product. BFLA: Any user can create products, not just admins.",
        "operationId": "createProduct",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the new product"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Description of the new product"
          },
          {
            "name": "price",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "description": "Price of the new product"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Category of the new product"
          }
        ],
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/products/with-stock": {
      "get": {
        "tags": [
          "Products",
          "Stock"
        ],
        "summary": "List all products with stock",
        "operationId": "listProductsWithStock",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "description": "Filter products based on user role (e.g., 'user'). This parameter seems to be sent by the UI.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "admin",
                "guest"
              ],
              "default": "user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of products with stock quantities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductWithStock"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{productId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product by product ID",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the product to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/api/products/{productId}/with-stock": {
      "get": {
        "tags": [
          "Products",
          "Stock"
        ],
        "summary": "Get product by product ID with stock",
        "operationId": "getProductByIdWithStock",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the product to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWithStock"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update product by product ID (Parameter Pollution & BFLA Target)",
        "description": "Updates a product. Parameter Pollution: `internal_status` can be set. BFLA: Any user can update products. Protected products cannot be modified.",
        "operationId": "updateProductById",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the product to update"
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "New name for the product"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "New description for the product"
          },
          {
            "name": "price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            },
            "description": "New price for the product"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "New category for the product"
          },
          {
            "name": "internal_status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "description": "Internal status (Parameter Pollution vulnerability)"
          }
        ],
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Action Forbidden: The product is protected for demo purposes and cannot be modified."
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete product by product ID (BFLA Target)",
        "description": "Deletes a product. BFLA: Any user can delete products. Protected products cannot be deleted.",
        "operationId": "deleteProductById",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the product to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Product deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Action Forbidden: Product is protected for demo purposes and cannot be deleted."
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/api/products/search": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Search for products by name (Injection Target)",
        "description": "Searches for products. The `name` parameter is a potential injection vector if not handled safely by the backend.",
        "operationId": "searchProducts",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search term for product name"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of matching products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (e.g., missing name parameter)"
          }
        }
      }
    },
    "/api/products/{productId}/stock": {
      "get": {
        "tags": [
          "Stock"
        ],
        "summary": "Get stock information for a product",
        "operationId": "getProductStock",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the product"
          }
        ],
        "responses": {
          "200": {
            "description": "Stock information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stock"
                }
              }
            }
          },
          "404": {
            "description": "Product not found or stock information not available"
          }
        }
      },
      "put": {
        "tags": [
          "Stock"
        ],
        "summary": "Update stock quantity for a product (BFLA Target)",
        "description": "Updates product stock. BFLA: Any user can update stock. Protected products have minimum stock level restrictions for direct updates.",
        "operationId": "updateProductStock",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the product"
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "New stock quantity"
          }
        ],
        "responses": {
          "200": {
            "description": "Stock updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stock"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (e.g., negative quantity)"
          },
          "403": {
            "description": "Action Forbidden: Product is protected and stock cannot be reduced below minimum, or insufficient privileges."
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "required": [
          "product_id",
          "name",
          "price"
        ],
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the product",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "Name of the product"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Description of the product"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Price of the product"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "description": "Category of the product"
          },
          "internal_status": {
            "type": "string",
            "nullable": true,
            "description": "Internal status of the product (demonstrates parameter pollution)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of product creation",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of last product update",
            "readOnly": true
          },
          "is_protected": {
            "type": "boolean",
            "description": "Indicates if this product is protected from deletion or certain updates",
            "default": false
          }
        }
      },
      "ProductWithStock": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Product"
          },
          {
            "type": "object",
            "required": [
              "stock_quantity"
            ],
            "properties": {
              "stock_quantity": {
                "type": "integer",
                "minimum": 0,
                "description": "Quantity of the product in stock"
              }
            }
          }
        ]
      },
      "Stock": {
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "description": "Foreign key to Product"
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity of the product in stock"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of last stock update",
            "readOnly": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}